# Refactor Expand/ExpandInner for bidirectional_astar