--- a/src/thor/bidirectional_astar.cc
+++ b/src/thor/bidirectional_astar.cc
@@ -135,108 +135,6 @@ void BidirectionalAStar::Init(const PointLL& origll, const PointLL& destll) {
   hierarchy_limits_reverse_ = costing_->GetHierarchyLimits();
 }
 
-// Returns true if function ended up adding an edge for expansion
-bool BidirectionalAStar::ExpandForward(GraphReader& graphreader,
-                                       const GraphId& node,
-                                       BDEdgeLabel& pred,
-                                       const uint32_t pred_idx,
-                                       const TimeInfo& time_info,
-                                       const bool invariant) {
-  // Get the tile and the node info. Skip if tile is null (can happen
-  // with regional data sets) or if no access at the node.
-  graph_tile_ptr tile = graphreader.GetGraphTile(node);
-  if (tile == nullptr) {
-    return false;
-  }
-  const NodeInfo* nodeinfo = tile->node(node);
-
-  // Keep track of superseded edges
-  uint32_t shortcuts = 0;
-
-  // Update the time information even if time is invariant to account for timezones
-  auto seconds_offset = invariant ? 0.f : pred.cost().secs;
-  auto offset_time = time_info.forward(seconds_offset, static_cast<int>(nodeinfo->timezone()));
-
-  // If we encounter a node with an access restriction like a barrier we allow a uturn
-  if (!costing_->Allowed(nodeinfo)) {
-    const DirectedEdge* opp_edge = nullptr;
-    const GraphId opp_edge_id = graphreader.GetOpposingEdgeId(pred.edgeid(), opp_edge, tile);
-    // Mark the predecessor as a deadend to be consistent with how the
-    // edgelabels are set when an *actual* deadend (i.e. some dangling OSM geometry)
-    // is labelled
-    pred.set_deadend(true);
-    // Check if edge is null before using it (can happen with regional data sets)
-    return opp_edge &&
-           ExpandForwardInner(graphreader, pred, nodeinfo, pred_idx,
-                              {opp_edge, opp_edge_id, edgestatus_forward_.GetPtr(opp_edge_id, tile)},
-                              shortcuts, tile, offset_time);
-  }
-
-  bool disable_uturn = false;
-  EdgeMetadata meta = EdgeMetadata::make(node, nodeinfo, tile, edgestatus_forward_);
-  EdgeMetadata uturn_meta{};
-
-  // Expand from end node in forward direction.
-  for (uint32_t i = 0; i < nodeinfo->edge_count(); ++i, ++meta) {
-
-    // Begin by checking if this is the opposing edge to pred.
-    // If so, it means we are attempting a u-turn. In that case, lets wait with evaluating
-    // this edge until last. If any other edges were emplaced, it means we should not
-    // even try to evaluate a u-turn since u-turns should only happen for deadends
-    uturn_meta = pred.opp_local_idx() == meta.edge->localedgeidx() ? meta : uturn_meta;
-
-    // Expand but only if this isnt the uturn, we'll try that later if nothing else works out
-    disable_uturn = (pred.opp_local_idx() != meta.edge->localedgeidx() &&
-                     ExpandForwardInner(graphreader, pred, nodeinfo, pred_idx, meta, shortcuts, tile,
-                                        offset_time)) ||
-                    disable_uturn;
-  }
-
-  // Handle transitions - expand from the end node of each transition
-  if (nodeinfo->transition_count() > 0) {
-    const NodeTransition* trans = tile->transition(nodeinfo->transition_index());
-    for (uint32_t i = 0; i < nodeinfo->transition_count(); ++i, ++trans) {
-      // if this is a downward transition (ups are always allowed) AND we are no longer allowed OR
-      // we cant get the tile at that level (local extracts could have this problem) THEN bail
-      graph_tile_ptr trans_tile = nullptr;
-      if ((!trans->up() && hierarchy_limits_forward_[trans->endnode().level()].StopExpanding()) ||
-          !(trans_tile = graphreader.GetGraphTile(trans->endnode()))) {
-        continue;
-      }
-      // setup for expansion at this level
-      hierarchy_limits_forward_[node.level()].up_transition_count += trans->up();
-      const auto* trans_node = trans_tile->node(trans->endnode());
-      EdgeMetadata trans_meta =
-          EdgeMetadata::make(trans->endnode(), trans_node, trans_tile, edgestatus_forward_);
-      uint32_t trans_shortcuts = 0;
-      // expand the edges from this node at this level
-      for (uint32_t i = 0; i < trans_node->edge_count(); ++i, ++trans_meta) {
-        disable_uturn = ExpandForwardInner(graphreader, pred, trans_node, pred_idx, trans_meta,
-                                           trans_shortcuts, trans_tile, offset_time) ||
-                        disable_uturn;
-      }
-    }
-  }
-
-  // Now, after having looked at all the edges, including edges on other levels,
-  // we can say if this is a deadend or not, and if so, evaluate the uturn-edge (if it exists)
-  if (!disable_uturn && uturn_meta) {
-    // If we found no suitable edge to add, it means we're at a deadend
-    // so lets go back and re-evaluate a potential u-turn
-    pred.set_deadend(true);
-
-    // TODO Is there a shortcut that supersedes our u-turn?
-    // Decide if we should expand a shortcut or the non-shortcut edge...
-
-    // Expand the uturn possiblity
-    disable_uturn = ExpandForwardInner(graphreader, pred, nodeinfo, pred_idx, uturn_meta, shortcuts,
-                                       tile, offset_time) ||
-                    disable_uturn;
-  }
-
-  return disable_uturn;
-}
-
 // Runs in the inner loop of `ExpandForward`, essentially evaluating if
 // the edge described in `meta` should be placed on the stack
 // as well as doing just that.
@@ -347,111 +245,6 @@ inline bool BidirectionalAStar::ExpandForwardInner(GraphReader& graphreader,
   return !(pred.not_thru_pruning() && meta.edge->not_thru());
 }
 
-// Expand from a node in reverse direction.
-//
-// Returns true if function ended up adding an edge for expansion
-bool BidirectionalAStar::ExpandReverse(GraphReader& graphreader,
-                                       const GraphId& node,
-                                       BDEdgeLabel& pred,
-                                       const uint32_t pred_idx,
-                                       const DirectedEdge* opp_pred_edge,
-                                       const TimeInfo& time_info,
-                                       const bool invariant) {
-  // Get the tile and the node info. Skip if tile is null (can happen
-  // with regional data sets) or if no access at the node.
-  graph_tile_ptr tile = graphreader.GetGraphTile(node);
-  if (tile == nullptr) {
-    return false;
-  }
-  const NodeInfo* nodeinfo = tile->node(node);
-
-  // Keep track of superseded edges
-  uint32_t shortcuts = 0;
-
-  // Update the time information even if time is invariant to account for timezones
-  auto seconds_offset = invariant ? 0.f : pred.cost().secs;
-  auto offset_time = time_info.reverse(seconds_offset, static_cast<int>(nodeinfo->timezone()));
-
-  // If we encounter a node with an access restriction like a barrier we allow a uturn
-  if (!costing_->Allowed(nodeinfo)) {
-    const DirectedEdge* opp_edge = nullptr;
-    const GraphId opp_edge_id = graphreader.GetOpposingEdgeId(pred.edgeid(), opp_edge, tile);
-    // Mark the predecessor as a deadend to be consistent with how the
-    // edgelabels are set when an *actual* deadend (i.e. some dangling OSM geometry)
-    // is labelled
-    pred.set_deadend(true);
-    // Check if edge is null before using it (can happen with regional data sets)
-    return opp_edge &&
-           ExpandReverseInner(graphreader, pred, opp_pred_edge, nodeinfo, pred_idx,
-                              {opp_edge, opp_edge_id, edgestatus_reverse_.GetPtr(opp_edge_id, tile)},
-                              shortcuts, tile, offset_time);
-  }
-
-  // We start off allowing uturns, and if we find any edge to expand from we disallow uturns here
-  bool disable_uturn = false;
-  EdgeMetadata meta = EdgeMetadata::make(node, nodeinfo, tile, edgestatus_reverse_);
-  EdgeMetadata uturn_meta{};
-
-  // Expand from end node in reverse direction.
-  for (uint32_t i = 0; i < nodeinfo->edge_count(); ++i, ++meta) {
-
-    // Begin by checking if this is the opposing edge to pred.
-    // If so, it means we are attempting a u-turn. In that case, lets wait with evaluating
-    // this edge until last. If any other edges were emplaced, it means we should not
-    // even try to evaluate a u-turn since u-turns should only happen for deadends
-    uturn_meta = pred.opp_local_idx() == meta.edge->localedgeidx() ? meta : uturn_meta;
-
-    // Expand but only if this isnt the uturn, we'll try that later if nothing else works out
-    disable_uturn = (pred.opp_local_idx() != meta.edge->localedgeidx() &&
-                     ExpandReverseInner(graphreader, pred, opp_pred_edge, nodeinfo, pred_idx, meta,
-                                        shortcuts, tile, offset_time)) ||
-                    disable_uturn;
-  }
-
-  // Handle transitions - expand from the end node of each transition
-  if (nodeinfo->transition_count() > 0) {
-    const NodeTransition* trans = tile->transition(nodeinfo->transition_index());
-    for (uint32_t i = 0; i < nodeinfo->transition_count(); ++i, ++trans) {
-      // if this is a downward transition (ups are always allowed) AND we are no longer allowed OR
-      // we cant get the tile at that level (local extracts could have this problem) THEN bail
-      graph_tile_ptr trans_tile = nullptr;
-      if ((!trans->up() && hierarchy_limits_reverse_[trans->endnode().level()].StopExpanding()) ||
-          !(trans_tile = graphreader.GetGraphTile(trans->endnode()))) {
-        continue;
-      }
-      // setup for expansion at this level
-      hierarchy_limits_reverse_[node.level()].up_transition_count += trans->up();
-      const auto* trans_node = trans_tile->node(trans->endnode());
-      EdgeMetadata trans_meta =
-          EdgeMetadata::make(trans->endnode(), trans_node, trans_tile, edgestatus_reverse_);
-      uint32_t trans_shortcuts = 0;
-      // expand the edges from this node at this level
-      for (uint32_t i = 0; i < trans_node->edge_count(); ++i, ++trans_meta) {
-        disable_uturn = ExpandReverseInner(graphreader, pred, opp_pred_edge, trans_node, pred_idx,
-                                           trans_meta, trans_shortcuts, trans_tile, offset_time) ||
-                        disable_uturn;
-      }
-    }
-  }
-
-  // Now, after having looked at all the edges, including edges on other levels,
-  // we can say if this is a deadend or not, and if so, evaluate the uturn-edge (if it exists)
-  if (!disable_uturn && uturn_meta) {
-    // If we found no suitable edge to add, it means we're at a deadend
-    // so lets go back and re-evaluate a potential u-turn
-    pred.set_deadend(true);
-
-    // TODO Is there a shortcut that supersedes our u-turn?
-    // Decide if we should expand a shortcut or the non-shortcut edge...
-
-    // We didn't add any shortcut of the uturn, therefore evaluate the regular uturn instead
-    disable_uturn = ExpandReverseInner(graphreader, pred, opp_pred_edge, nodeinfo, pred_idx,
-                                       uturn_meta, shortcuts, tile, offset_time) ||
-                    disable_uturn;
-  }
-
-  return disable_uturn;
-}
 // Runs in the inner loop of `ExpandReverse`, essentially evaluating if
 // the edge described in `meta` should be placed on the stack
 // as well as doing just that.
@@ -767,8 +560,8 @@ BidirectionalAStar::GetBestPath(valhalla::Location& origin,
       }
 
       // Expand from the end node in forward direction.
-      ExpandForward(graphreader, fwd_pred.endnode(), fwd_pred, forward_pred_idx, forward_time_info,
-                    invariant);
+      Expand<ExpansionType::forward>(graphreader, fwd_pred.endnode(), fwd_pred, forward_pred_idx,
+                                     nullptr, forward_time_info, invariant);
     } else {
       // Expand reverse - set to get next edge from reverse adj. list on the next pass
       expand_forward = false;
@@ -791,8 +584,8 @@ BidirectionalAStar::GetBestPath(valhalla::Location& origin,
           graphreader.GetGraphTile(rev_pred.opp_edgeid())->directededge(rev_pred.opp_edgeid());
 
       // Expand from the end node in reverse direction.
-      ExpandReverse(graphreader, rev_pred.endnode(), rev_pred, reverse_pred_idx, opp_pred_edge,
-                    reverse_time_info, invariant);
+      Expand<ExpansionType::reverse>(graphreader, rev_pred.endnode(), rev_pred, reverse_pred_idx,
+                                     opp_pred_edge, reverse_time_info, invariant);
     }
   }
   return {}; // If we are here the route failed