# Issue

Similar in vein to https://github.com/valhalla/valhalla/pull/2969 <- moves duplicated code into template functions.

This is just the first step toward consolidating as much of this edge traversal code as possible.  At this stage, I expect no behavioural changes.

## Tasklist

 - [ ] Add tests
 - [ ] Add #fixes with the issue number that this PR addresses
 - [ ] Update the docs with any new request parameters or changes to behavior described
 - [x] Update the [changelog](CHANGELOG.md)

## Requirements / Relations

 Link any requirements here. Other pull requests this PR is based on?