@ktatterso 2021-03-29
You mentioned that by consolidating the forward/reverse algo's you fixed some issues?  May I ask what lines of code were different in forward/reverse that are now corrected in your implementation?  thx