@danpat 2021-04-02
Benchmark results are showing a notable slowdown here (nearly 10%, using `bench/thor/routes.cc`), for routing:

Before:
```
----------------------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------
BM_UtrechtBidirectionalAstar_mean         14.9 ms         14.9 ms           10 Routes=20.6789k/s
BM_UtrechtBidirectionalAstar_median       14.9 ms         14.9 ms           10 Routes=20.6684k/s
BM_UtrechtBidirectionalAstar_stddev      0.331 ms        0.328 ms           10 Routes=455.241/s
BM_GetSpeed_mean                          6.29 ns         6.29 ns           10
BM_GetSpeed_median                        6.31 ns         6.31 ns           10
BM_GetSpeed_stddev                       0.147 ns        0.145 ns           10
BM_Sif_Allowed_mean                       4.59 ns         4.58 ns           10
BM_Sif_Allowed_median                     4.59 ns         4.58 ns           10
BM_Sif_Allowed_stddev                    0.090 ns        0.089 ns           10
```

After:
```
----------------------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------

BM_UtrechtBidirectionalAstar_mean         15.7 ms         15.7 ms           10 Routes=17.8932k/s
BM_UtrechtBidirectionalAstar_median       15.6 ms         15.6 ms           10 Routes=17.9881k/s
BM_UtrechtBidirectionalAstar_stddev      0.484 ms        0.481 ms           10 Routes=541.108/s
BM_GetSpeed_mean                          6.28 ns         6.28 ns           10
BM_GetSpeed_median                        6.31 ns         6.31 ns           10
BM_GetSpeed_stddev                       0.091 ns        0.091 ns           10
BM_Sif_Allowed_mean                       4.36 ns         4.35 ns           10
BM_Sif_Allowed_median                     4.34 ns         4.34 ns           10
BM_Sif_Allowed_stddev                    0.121 ns        0.120 ns           10
```