@danpat 2021-04-04
Hmm, ok - I'm now not convinced this wasn't noise caused by other processes on my system.  I killed of everything unnecessary, and re-ran the following:

```
$ ./bench/thor/benchmark-routes  --benchmark_repetitions=50 --benchmark_filter=BM_UtrechtBidirectionalAstar
```

`master` branch:

```
----------------------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------
BM_UtrechtBidirectionalAstar              17.2 ms         17.2 ms           42 Routes=17.117k/s
BM_UtrechtBidirectionalAstar              17.4 ms         17.2 ms           42 Routes=17.0869k/s
BM_UtrechtBidirectionalAstar              15.6 ms         15.5 ms           42 Routes=18.9159k/s
BM_UtrechtBidirectionalAstar              19.9 ms         19.8 ms           42 Routes=14.832k/s
BM_UtrechtBidirectionalAstar              19.6 ms         19.5 ms           42 Routes=15.0758k/s
BM_UtrechtBidirectionalAstar              19.6 ms         19.6 ms           42 Routes=14.9891k/s
BM_UtrechtBidirectionalAstar              18.9 ms         18.9 ms           42 Routes=15.569k/s
BM_UtrechtBidirectionalAstar              17.2 ms         17.2 ms           42 Routes=17.1399k/s
BM_UtrechtBidirectionalAstar              16.0 ms         16.0 ms           42 Routes=18.3816k/s
BM_UtrechtBidirectionalAstar              16.8 ms         16.8 ms           42 Routes=17.5237k/s
BM_UtrechtBidirectionalAstar              15.6 ms         15.6 ms           42 Routes=18.808k/s
BM_UtrechtBidirectionalAstar              16.2 ms         16.2 ms           42 Routes=18.1592k/s
BM_UtrechtBidirectionalAstar              14.7 ms         14.7 ms           42 Routes=20.0559k/s
BM_UtrechtBidirectionalAstar              14.9 ms         14.8 ms           42 Routes=19.8002k/s
BM_UtrechtBidirectionalAstar              15.3 ms         15.3 ms           42 Routes=19.249k/s
BM_UtrechtBidirectionalAstar              16.0 ms         16.0 ms           42 Routes=18.3237k/s
BM_UtrechtBidirectionalAstar              14.9 ms         14.9 ms           42 Routes=19.6799k/s
BM_UtrechtBidirectionalAstar              14.5 ms         14.5 ms           42 Routes=20.2517k/s
BM_UtrechtBidirectionalAstar              15.0 ms         15.0 ms           42 Routes=19.6139k/s
BM_UtrechtBidirectionalAstar              16.7 ms         16.7 ms           42 Routes=17.6493k/s
BM_UtrechtBidirectionalAstar              14.7 ms         14.7 ms           42 Routes=19.9754k/s
BM_UtrechtBidirectionalAstar              15.0 ms         15.0 ms           42 Routes=19.5822k/s
BM_UtrechtBidirectionalAstar              15.1 ms         15.1 ms           42 Routes=19.4817k/s
BM_UtrechtBidirectionalAstar              15.3 ms         15.3 ms           42 Routes=19.2209k/s
BM_UtrechtBidirectionalAstar              15.5 ms         15.5 ms           42 Routes=19.0147k/s
BM_UtrechtBidirectionalAstar              14.8 ms         14.8 ms           42 Routes=19.8693k/s
BM_UtrechtBidirectionalAstar              15.8 ms         15.8 ms           42 Routes=18.6612k/s
BM_UtrechtBidirectionalAstar              15.5 ms         15.5 ms           42 Routes=18.9962k/s
BM_UtrechtBidirectionalAstar              14.8 ms         14.8 ms           42 Routes=19.8741k/s
BM_UtrechtBidirectionalAstar              15.4 ms         15.4 ms           42 Routes=19.0722k/s
BM_UtrechtBidirectionalAstar              15.5 ms         15.5 ms           42 Routes=18.9426k/s
BM_UtrechtBidirectionalAstar              15.3 ms         15.3 ms           42 Routes=19.1841k/s
BM_UtrechtBidirectionalAstar              15.1 ms         15.1 ms           42 Routes=19.4529k/s
BM_UtrechtBidirectionalAstar              14.5 ms         14.5 ms           42 Routes=20.3203k/s
BM_UtrechtBidirectionalAstar              15.1 ms         15.1 ms           42 Routes=19.5227k/s
BM_UtrechtBidirectionalAstar              14.6 ms         14.6 ms           42 Routes=20.0989k/s
BM_UtrechtBidirectionalAstar              14.7 ms         14.7 ms           42 Routes=20.052k/s
BM_UtrechtBidirectionalAstar              14.9 ms         14.9 ms           42 Routes=19.674k/s
BM_UtrechtBidirectionalAstar              15.0 ms         15.0 ms           42 Routes=19.6388k/s
BM_UtrechtBidirectionalAstar              14.8 ms         14.8 ms           42 Routes=19.8346k/s
BM_UtrechtBidirectionalAstar              14.5 ms         14.5 ms           42 Routes=20.2716k/s
BM_UtrechtBidirectionalAstar              14.9 ms         14.9 ms           42 Routes=19.7442k/s
BM_UtrechtBidirectionalAstar              15.5 ms         15.4 ms           42 Routes=19.0386k/s
BM_UtrechtBidirectionalAstar              14.6 ms         14.6 ms           42 Routes=20.1851k/s
BM_UtrechtBidirectionalAstar              14.8 ms         14.8 ms           42 Routes=19.8181k/s
BM_UtrechtBidirectionalAstar              14.8 ms         14.8 ms           42 Routes=19.9072k/s
BM_UtrechtBidirectionalAstar              15.0 ms         15.0 ms           42 Routes=19.653k/s
BM_UtrechtBidirectionalAstar              14.8 ms         14.8 ms           42 Routes=19.8744k/s
BM_UtrechtBidirectionalAstar              14.6 ms         14.6 ms           42 Routes=20.0962k/s
BM_UtrechtBidirectionalAstar              14.9 ms         14.9 ms           42 Routes=19.7063k/s
BM_UtrechtBidirectionalAstar_mean         15.6 ms         15.6 ms           50 Routes=18.9397k/s
BM_UtrechtBidirectionalAstar_median       15.1 ms         15.1 ms           50 Routes=19.5022k/s
BM_UtrechtBidirectionalAstar_stddev       1.36 ms         1.35 ms           50 Routes=1.41591k/s
```

This branch:
```
----------------------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------
BM_UtrechtBidirectionalAstar              15.6 ms         15.6 ms           42 Routes=18.8363k/s
BM_UtrechtBidirectionalAstar              15.0 ms         15.0 ms           42 Routes=19.5564k/s
BM_UtrechtBidirectionalAstar              14.8 ms         14.8 ms           42 Routes=19.8808k/s
BM_UtrechtBidirectionalAstar              15.4 ms         15.4 ms           42 Routes=19.1489k/s
BM_UtrechtBidirectionalAstar              15.7 ms         15.7 ms           42 Routes=18.7539k/s
BM_UtrechtBidirectionalAstar              16.2 ms         16.2 ms           42 Routes=18.1659k/s
BM_UtrechtBidirectionalAstar              15.6 ms         15.6 ms           42 Routes=18.8739k/s
BM_UtrechtBidirectionalAstar              14.8 ms         14.8 ms           42 Routes=19.8616k/s
BM_UtrechtBidirectionalAstar              15.3 ms         15.3 ms           42 Routes=19.2233k/s
BM_UtrechtBidirectionalAstar              14.8 ms         14.8 ms           42 Routes=19.8223k/s
BM_UtrechtBidirectionalAstar              15.2 ms         15.2 ms           42 Routes=19.3204k/s
BM_UtrechtBidirectionalAstar              15.6 ms         15.6 ms           42 Routes=18.8202k/s
BM_UtrechtBidirectionalAstar              15.6 ms         15.6 ms           42 Routes=18.898k/s
BM_UtrechtBidirectionalAstar              15.2 ms         15.2 ms           42 Routes=19.2952k/s
BM_UtrechtBidirectionalAstar              16.4 ms         16.4 ms           42 Routes=17.9377k/s
BM_UtrechtBidirectionalAstar              14.7 ms         14.7 ms           42 Routes=19.96k/s
BM_UtrechtBidirectionalAstar              15.5 ms         15.5 ms           42 Routes=18.9899k/s
BM_UtrechtBidirectionalAstar              14.8 ms         14.8 ms           42 Routes=19.9057k/s
BM_UtrechtBidirectionalAstar              15.1 ms         15.1 ms           42 Routes=19.5063k/s
BM_UtrechtBidirectionalAstar              16.1 ms         16.0 ms           42 Routes=18.3212k/s
BM_UtrechtBidirectionalAstar              15.2 ms         15.2 ms           42 Routes=19.3172k/s
BM_UtrechtBidirectionalAstar              15.6 ms         15.6 ms           42 Routes=18.9067k/s
BM_UtrechtBidirectionalAstar              15.6 ms         15.6 ms           42 Routes=18.8519k/s
BM_UtrechtBidirectionalAstar              15.0 ms         15.0 ms           42 Routes=19.5903k/s
BM_UtrechtBidirectionalAstar              15.2 ms         15.2 ms           42 Routes=19.3551k/s
BM_UtrechtBidirectionalAstar              15.2 ms         15.2 ms           42 Routes=19.3897k/s
BM_UtrechtBidirectionalAstar              16.0 ms         16.0 ms           42 Routes=18.4274k/s
BM_UtrechtBidirectionalAstar              16.0 ms         16.0 ms           42 Routes=18.4082k/s
BM_UtrechtBidirectionalAstar              14.8 ms         14.8 ms           42 Routes=19.9011k/s
BM_UtrechtBidirectionalAstar              15.5 ms         15.5 ms           42 Routes=19.0067k/s
BM_UtrechtBidirectionalAstar              14.9 ms         14.9 ms           42 Routes=19.6923k/s
BM_UtrechtBidirectionalAstar              15.8 ms         15.8 ms           42 Routes=18.6367k/s
BM_UtrechtBidirectionalAstar              15.6 ms         15.6 ms           42 Routes=18.8815k/s
BM_UtrechtBidirectionalAstar              14.8 ms         14.8 ms           42 Routes=19.8339k/s
BM_UtrechtBidirectionalAstar              15.9 ms         15.9 ms           42 Routes=18.5322k/s
BM_UtrechtBidirectionalAstar              15.9 ms         15.9 ms           42 Routes=18.5239k/s
BM_UtrechtBidirectionalAstar              14.8 ms         14.8 ms           42 Routes=19.8834k/s
BM_UtrechtBidirectionalAstar              15.5 ms         15.5 ms           42 Routes=18.9722k/s
BM_UtrechtBidirectionalAstar              15.0 ms         15.0 ms           42 Routes=19.6397k/s
BM_UtrechtBidirectionalAstar              15.1 ms         15.1 ms           42 Routes=19.4356k/s
BM_UtrechtBidirectionalAstar              17.3 ms         17.3 ms           42 Routes=16.9957k/s
BM_UtrechtBidirectionalAstar              17.6 ms         17.4 ms           42 Routes=16.8671k/s
BM_UtrechtBidirectionalAstar              16.5 ms         16.5 ms           42 Routes=17.8646k/s
BM_UtrechtBidirectionalAstar              15.8 ms         15.8 ms           42 Routes=18.6304k/s
BM_UtrechtBidirectionalAstar              15.0 ms         15.0 ms           42 Routes=19.64k/s
BM_UtrechtBidirectionalAstar              15.9 ms         15.9 ms           42 Routes=18.4912k/s
BM_UtrechtBidirectionalAstar              15.6 ms         15.6 ms           42 Routes=18.827k/s
BM_UtrechtBidirectionalAstar              15.1 ms         15.1 ms           42 Routes=19.4265k/s
BM_UtrechtBidirectionalAstar              15.8 ms         15.8 ms           42 Routes=18.6015k/s
BM_UtrechtBidirectionalAstar              15.5 ms         15.5 ms           42 Routes=18.9775k/s
BM_UtrechtBidirectionalAstar_mean         15.5 ms         15.5 ms           50 Routes=19.0117k/s
BM_UtrechtBidirectionalAstar_median       15.5 ms         15.5 ms           50 Routes=18.9837k/s
BM_UtrechtBidirectionalAstar_stddev      0.600 ms        0.589 ms           50 Routes=694.207/s
```

However, even between runs, I'm seeing variability from 15k routes/sec to 19k routes/sec on `master` alone - so my benchmarking system does not appear to be very stable.  I'm going to look and see if I can somehow better control the processing environment.