@danpat 2021-04-07
@kevinkreiser ah, you're right - the `benchmark-routes` was building a new `BidirectionalAstar` object on each test iteration - moving that outside the test loop has stabilized the results:

### master
```
$ for i in 1 2 3 4 5 ; do ./bench/thor/benchmark-routes  --benchmark_min_time=3 --benchmark_filter=BM_UtrechtBidirectionalAstar --benchmark_counters_tabular=true --benchmark_repetitions=3 2>&1 | grep _m ; done
-----------------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations     Routes
-----------------------------------------------------------------------------------------
BM_UtrechtBidirectionalAstar_mean         13.1 ms         13.0 ms            3     2.275k
BM_UtrechtBidirectionalAstar_median       13.1 ms         13.1 ms            3     2.275k
BM_UtrechtBidirectionalAstar_mean         13.0 ms         13.0 ms            3     2.261k
BM_UtrechtBidirectionalAstar_median       13.0 ms         13.0 ms            3     2.261k
BM_UtrechtBidirectionalAstar_mean         13.0 ms         13.0 ms            3      2.24k
BM_UtrechtBidirectionalAstar_median       13.0 ms         13.0 ms            3      2.24k
BM_UtrechtBidirectionalAstar_mean         13.0 ms         13.0 ms            3     2.268k
BM_UtrechtBidirectionalAstar_median       13.0 ms         13.0 ms            3     2.268k
BM_UtrechtBidirectionalAstar_mean         13.0 ms         12.9 ms            3     2.261k
BM_UtrechtBidirectionalAstar_median       13.0 ms         13.0 ms            3     2.261k
```

### this branch
```
-----------------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations     Routes
-----------------------------------------------------------------------------------------
BM_UtrechtBidirectionalAstar_mean         13.3 ms         13.3 ms            3     2.198k
BM_UtrechtBidirectionalAstar_median       13.2 ms         13.2 ms            3     2.198k
BM_UtrechtBidirectionalAstar_mean         13.3 ms         13.2 ms            3     2.219k
BM_UtrechtBidirectionalAstar_median       13.2 ms         13.2 ms            3     2.219k
BM_UtrechtBidirectionalAstar_mean         13.2 ms         13.2 ms            3     2.233k
BM_UtrechtBidirectionalAstar_median       13.2 ms         13.2 ms            3     2.233k
BM_UtrechtBidirectionalAstar_mean         13.1 ms         13.1 ms            3     2.212k
BM_UtrechtBidirectionalAstar_median       13.2 ms         13.2 ms            3     2.212k
BM_UtrechtBidirectionalAstar_mean         13.2 ms         13.2 ms            3     2.072k
BM_UtrechtBidirectionalAstar_median       13.2 ms         13.2 ms            3     2.072k
```

So it appears something in the refactor has introduced about 100us of overhead for this test case.

My vote is that <1% hit here is likely worth the reduced code complexity - and can likely be addressed by some close analysis of the generated code.  I re-added the inline hint, but the difference still remains.