@danpat 2021-04-07
Yeah, that's fair.  I grabbed the first 33 routes from `test_requests/random.txt`, and used a recent whole-planet extract.  Here is the length/distribution of the tested routes:

<img width="811" alt="Screen Shot 2021-04-07 at 9 09 54 AM" src="https://user-images.githubusercontent.com/1892250/113898771-07100d80-9781-11eb-949e-f3c380653740.png">

I ran the benchmarks a few times to ensure all the caches were warmed up.  To understand the data below: it takes about 28 seconds to run all 33 routes single-threaded, one after the other.  Each benchmark run did this 3 times in a row within the same process, but with a new `BidirectionalAStar` object each time (setup time for that was outside the timing loop):

### master
```
-----------------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations     Routes
-----------------------------------------------------------------------------------------
BM_UtrechtBidirectionalAstar             28394 ms        28005 ms            1         33
BM_UtrechtBidirectionalAstar             26940 ms        26841 ms            1         33
BM_UtrechtBidirectionalAstar             28469 ms        28438 ms            1         33
BM_UtrechtBidirectionalAstar_mean        27934 ms        27761 ms            3         33
BM_UtrechtBidirectionalAstar_median      28394 ms        28005 ms            3         33
BM_UtrechtBidirectionalAstar_stddev        862 ms          826 ms            3          0
```

### this branch
```
-----------------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations     Routes
-----------------------------------------------------------------------------------------
BM_UtrechtBidirectionalAstar             28408 ms        28381 ms            1         33
BM_UtrechtBidirectionalAstar             28076 ms        28047 ms            1         33
BM_UtrechtBidirectionalAstar             26210 ms        26187 ms            1         33
BM_UtrechtBidirectionalAstar_mean        27565 ms        27538 ms            3         33
BM_UtrechtBidirectionalAstar_median      28076 ms        28047 ms            3         33
BM_UtrechtBidirectionalAstar_stddev       1185 ms         1182 ms            3          0
```